requirements.txt
requests~=2.32.5
pydantic~=2.11.7
pytest~=8.4.1
python-dotenv~=1.1.1
Faker~=37.5.3
allure-pytest~=2.15.0